Moved scenarios over into main factory system.
