+++ docs/api/common/factories/index.md
@@ -188,3 +188,44 @@ identical, report = linked_testkit.diff_results(
 
 assert identical, report
 ```
+
+## Testing with scenarios
+
+For more complex integration tests, the factory provides a scenario system. This allows you to stand up a fully-populated backend with a single context manager, [`setup_scenario()`][matchbox.common.factories.scenarios.setup_scenario]. This is particularly useful for testing database adapters and end-to-end methodologies.
+
+The main usage pattern is to call `setup_scenario()` with a backend adapter and a named scenario. The context manager yields a `TestkitDAG` containing all the sources and models created for the scenario, giving you access to the ground truth.
+
+```python
+from matchbox.common.factories import setup_scenario
+
+def test_my_adapter_function(my_backend_adapter):
+    with setup_scenario(my_backend_adapter, "link") as dag:
+        # The backend is now populated with the 'link' scenario
+        # dag.sources contains the source testkits
+        # dag.models contains the model testkits
+        
+        # Now you can call the function you want to test
+        results = my_backend_adapter.query(resolution="final_join")
+        
+        # You can use the dag to verify the results
+        assert len(results) > 0
+
+```
+
+The scenario system is cached, so subsequent runs of the same scenario are significantly faster.
+
+### Available scenarios
+
+The following built-in scenarios are available. They are built on top of each other, so `link` includes all the steps from `dedupe`, which includes `index`, and so on.
+
+*   **`bare`**: Creates a set of linked sources and writes them to the data warehouse, but does not interact with the matchbox backend.
+*   **`index`**: Takes the `bare` scenario and indexes all the sources in the matchbox backend.
+*   **`dedupe`**: Takes the `index` scenario and adds perfectly deduplicated models for each source.
+*   **`probabilistic_dedupe`**: Like `dedupe`, but the models produce probabilistic scores rather than perfect matches.
+*   **`link`**: Takes the `dedupe` scenario and adds linking models between the deduplicated sources, culminating in a `final_join` resolution.
+*   **`alt_dedupe`**: A specialised scenario with two alternative deduplication models for the same source.
+*   **`convergent`**: A specialised scenario where two different sources index to almost identical data.
+
+### Creating new scenarios
+
+You can create your own scenarios by writing a builder function and registering it with the [`@register_scenario`][matchbox.common.factories.scenarios.register_scenario] decorator. This allows you to build reusable, complex data setups for your tests.
