+++ src/matchbox/common/factories/__init__.py
@@ -1,1 +1,15 @@
 """Factory functions for the testkit system."""
+
+from matchbox.common.factories.scenarios import (
+    SCENARIO_REGISTRY,
+    DevelopmentSettings,
+    ScenarioBuilder,
+    setup_scenario,
+)
+
+__all__ = [
+    "SCENARIO_REGISTRY",
+    "DevelopmentSettings",
+    "ScenarioBuilder",
+    "setup_scenario",
+]
